Extract `run` into a helper function
